requirements.txt
antlr4-python3-runtime==4.8
beautifulsoup4==4.11.1
blis==0.4.1
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cymem==2.0.6
Cython==0.29.28
dataclasses==0.6
editdistance==0.6.0
fairseq==0.10.0
fastBPE==0.1.0
filelock==3.6.0
Flask==2.2.2
huggingface-hub==0.5.1
hydra-core==1.1.2
idna==3.3
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
mysql-connector-python==8.0.29
numpy==1.22.3
omegaconf==2.1.2
packaging==21.3
pathy==0.6.1
plac==1.1.3
portalocker==2.4.0
preshed==3.0.6
protobuf==3.20.1
pycparser==2.21
pydantic==1.8.2
pymediawiki==0.7.1
pyparsing==3.0.8
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacrebleu==2.0.0
sacremoses==0.0.49
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==2.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==1.0.5
tabulate==0.8.9
thinc==7.4.1
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
vncorenlp==1.0.3
wasabi==0.9.1
Werkzeug==2.2.2
zipp==3.11.0
